% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Weighted Leverating}
\usage{
algo_leverage(X, Y, r, method = "uniform")
}
\arguments{
\item{X}{A matrix of input values}

\item{Y}{A vector of outputs}

\item{r}{The subsample size}

\item{method}{Either 'uniform' or 'leverage' methods. The uniform method}
}
\value{
Returns the coefficients computed using subsampling by selecting samples either uniformly or with probabilities weighted by the leverage scores.
}
\description{
A method for computing the coefficients in a linear regression using subsampling by selecting samples either uniformly or with probabilities weighted by the leverage scores.
}
