% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Net}
\usage{
elnet_coord(X, Y, lambdas = seq(0.01, 3, length.out = 25), alpha = 0.5)
}
\arguments{
\item{X}{Matrix of inputs}

\item{Y}{Vector of outputs}

\item{lambdas}{A sequence of lambda values for the elastic net algorithm}

\item{alpha}{The elastic net mixing parameter. Alpha should be between 0 and 1}
}
\value{
A matrix of the fitted coefficients. The coefficients in column i correspond to lambda_i
}
\description{
Fit a linear model with an elastic net penalty for a sequence of values for the regularization coefficient lambda.
}
