% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Ordinary Least Squares}
\usage{
solve_ols(A, b, method = "Gauss", numcores = 1, niter = 10000)
}
\arguments{
\item{A}{A matrix of coefficients}

\item{b}{A vector}

\item{method}{Either "Gauss" for the Gauss-Seidel method or "Jacobi" for the Jacobi Method. Implements the parallel-Jacobi method if numcores is more than 1 and the sequential Jacobi method if numcores is 1.}

\item{numcores}{An integer for the numebr of cores to use for computing the Jacobi method}

\item{niter}{The number of iterations to perform.}
}
\value{
A solution x to the equation Ax = b
}
\description{
Computes the Ordinary Least Squares solution using either the Gauss-Seidel or the Jacobi iterative methods.
}
